+++ src/main/java/com/infoshare/alpha/wwr/servlet/FacilityServlet.java
@@ -1,6 +1,6 @@
 package com.infoshare.alpha.wwr.servlet;
 
-import com.infoshare.alpha.wwr.domain.facilities.datastorage.FacilitiesJsonStorage;
+import com.infoshare.alpha.wwr.domain.facilities.readmodel.FacilitiesReadModelDb;
 
 import javax.inject.Inject;
 import javax.servlet.ServletException;
@@ -13,12 +13,12 @@ import java.io.IOException;
 public class FacilityServlet extends BaseWwrServlet {
 
     @Inject
-    FacilitiesJsonStorage facilitiesJsonStorage;
+    FacilitiesReadModelDb facilitiesReadModelDb;
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         resp.getWriter().println("Hello world");
-        if (facilitiesJsonStorage.load().getFacilities().isEmpty()){
+        if (facilitiesReadModelDb.getAll().getFacilities().isEmpty()){
             resp.getWriter().println("error!");
         } else {
             resp.getWriter().println("success!");
