+++ src/main/java/com/infoshare/alpha/wwr/utils/InputForms.java
@@ -2,7 +2,6 @@ package com.infoshare.alpha.wwr.utils;
 
 import com.infoshare.alpha.wwr.common.Address;
 import com.infoshare.alpha.wwr.common.Pesel;
-import com.infoshare.alpha.wwr.common.PeselException;
 import com.infoshare.alpha.wwr.common.Service;
 import com.infoshare.alpha.wwr.common.Facility;
 import com.infoshare.alpha.wwr.domain.patients.entity.Parent;
